Move transaction info pager adapter into a separate file
